backend/requirements.txt
# FastAPI and related
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-decouple==3.8

# OAuth
authlib==1.2.1
httpx==0.25.2
requests==2.31.0

# CORS for frontend communication
fastapi-cors==0.0.6

# Environment variables
python-dotenv==1.0.0

# Web3 & Wallet Authentication
eth-account==0.10.0

# Testing
pytest==8.2.0
pytest-asyncio==0.24.0
httpx==0.25.2
pytest-cov==4.1.0